skin-model/requirements.txt
# Runtime dependencies for skin-model Flask inference service
# Note: torch/torchvision are intentionally not pinned here because your
# high-performance server already has a prepared PyTorch 3.11 environment.

flask>=3.0,<4.0
pillow>=10.0
numpy>=1.24
timm>=1.0.0
open-clip-torch>=2.24.0

